+++ src/player.cpp
@@ -121,5 +121,5 @@ void Player::Move()
 }
 Rectangle Player::GetHitBoxLoc()
 {
-	return {position.x+5, position.y+ 25, width-10, 15};
+	return {position.x+5, position.y+ 25, static_cast<float>(width-10), 15};
 }
